# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DVDThèque",
        "description": "Documentation de l'API DVDThèque",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "d618a255475240ba3d40cc39692cd924",
                "requestBody": {
                    "description": "The Token Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return access token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Access token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The provided credentials are incorrect"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "operationId": "3604fee5a25e84fa9437aed01a34e157",
                "responses": {
                    "200": {
                        "description": "Get auth user informations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Auth"
                ],
                "operationId": "271c4df7c5cfd126746f7a243e859273",
                "requestBody": {
                    "description": "The user informations",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test user 2"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "example": "password2"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "example": "password2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated user informations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/boxes": {
            "post": {
                "tags": [
                    "Boxes"
                ],
                "operationId": "d86184063a3a87db8ccf49c9025ac1d2",
                "requestBody": {
                    "description": "The Token Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "bar_code": {
                                        "type": "string",
                                        "example": "3512392506697"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get box informations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Box"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Box not found"
                    },
                    "422": {
                        "description": "Bad informations"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/boxes/{id}": {
            "get": {
                "tags": [
                    "Boxes"
                ],
                "operationId": "5b5e0e131877a6bbaeadbafe3bf5870d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get box informations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Box"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Box not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/me/boxes": {
            "get": {
                "tags": [
                    "Boxes"
                ],
                "operationId": "eb4eeac0a1a05dba773da6791b2350c7",
                "parameters": [
                    {
                        "name": "wishlist",
                        "in": "query",
                        "example": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get boxes associated to auth user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LightBox"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string"
                                                },
                                                "next": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad informations"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/me/boxes/{id}": {
            "post": {
                "tags": [
                    "Boxes"
                ],
                "operationId": "69eb09530683c75c0bc4fb2a1e575441",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "wishlist": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get box informations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Box"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Box not found"
                    },
                    "422": {
                        "description": "Bad informations"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Boxes"
                ],
                "operationId": "42c5a0179f23cfd42840b3257eff5acd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Box removed from user list"
                    },
                    "404": {
                        "description": "Box not found"
                    },
                    "422": {
                        "description": "Bad informations"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Box": {
                "title": "Box",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "original_title": {
                        "type": "string"
                    },
                    "year": {
                        "type": "integer"
                    },
                    "synopsis": {
                        "type": "string"
                    },
                    "edition": {
                        "type": "string"
                    },
                    "editor": {
                        "type": "string"
                    },
                    "illustration": {
                        "$ref": "#/components/schemas/Illustration"
                    },
                    "boxes": {
                        "type": "array",
                        "items": {
                            "description": "Box Object",
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "Illustration": {
                "title": "Illustration",
                "properties": {
                    "original": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LightBox": {
                "title": "LightBox",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "illustration": {
                        "$ref": "#/components/schemas/Illustration"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}